requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
pyrsistent==0.19.3
python-decouple==3.8
pytz==2023.3
SQLAlchemy==2.0.15
typing_extensions==4.6.3
Werkzeug==2.3.4
